+++ src/features/settings/store/reducer.ts
@@ -1,7 +1,7 @@
 import { createSlice, type PayloadAction } from '@reduxjs/toolkit'
 import { events } from '../../gameboard/events'
 
-import { HypnoMode, PlayerGender, PlayerParts, PornQuality, type Credentials, type EventToken, type PornList } from '../../gameboard/types'
+import { HypnoMode, PlayerGender, PlayerParts, PornQuality, PornService, type Credentials, type EventToken, type PornList } from '../../gameboard/types'
 
 export interface ISettingsState {
   dialogShown: boolean
@@ -12,7 +12,7 @@ export interface ISettingsState {
   steepness: number
   warmpupDuration: number
   duration: number
-  credentials?: Credentials
+  credentials: Credentials
   porn: PornList
   pornToCumTo: PornList
   pornQuality: PornQuality
@@ -40,7 +40,7 @@ const settingsSlice = createSlice({
     steepness: 0.05,
     warmpupDuration: 1800,
     duration: 6000,
-    credentials: undefined,
+    credentials: {},
     porn: [],
     pornToCumTo: [],
     pornQuality: PornQuality.LOW,
@@ -75,8 +75,17 @@ const settingsSlice = createSlice({
     SetDuration: (state, action: PayloadAction<number>) => {
       state.duration = action.payload
     },
-    SetCredentials: (state, action: PayloadAction<Credentials | undefined>) => {
-      state.credentials = action.payload
+    SetAllCredentials: (state, action: PayloadAction<Credentials>) => {
+      state.credentials = action.payload;
+    },
+    SetCredentials: (state, action: PayloadAction<{service: PornService, credentials: Credentials[PornService] | undefined}>) => {
+      if (action.payload.service === PornService.E621) {
+        state.credentials[PornService.E621] = action.payload.credentials as Credentials[PornService.E621]
+      }
+
+      if (action.payload.service === PornService.REDGIFS) {
+        state.credentials[PornService.REDGIFS] = action.payload.credentials as Credentials[PornService.REDGIFS]
+      }
     },
     SetPornList: (state, action: PayloadAction<PornList>) => {
       state.porn = action.payload